requirements.txt
requests==2.19.1
PyMySQL==0.9.2
PyYAML==3.13
